requirements.txt
crewai
crewai-tools
python-dotenv
yfinance
requests
pandas
flask
